+++ crates/cubecl-runtime/src/throughput/mod.rs
@@ -1,5 +1,9 @@
 mod base;
+mod benchmarker;
+mod cache;
 mod cmma;
 
 pub use base::*;
+pub use benchmarker::*;
+pub use cache::*;
 pub use cmma::*;
